vision/vit/pytorch/requirements.txt
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
transformers==4.18.0
pyyaml==5.4.1
protobuf==3.19.4
wandb==0.12.8
pytest==6.2.5
pytest-pythonpath==0.7.4
torchvision==0.14.1+cpu; python_version > '3.6'
torchvision==0.11.0+cpu; python_version <= '3.6'
pillow==8.4.0
attrdict==2.0.1
horovod==0.24.0
randaugment==1.0.2
numpy==1.23.5; python_version > '3.6'
numpy==1.19.5; python_version <= '3.6'
git+https://github.com/graphcore/examples-utils@latest_stable
